Fixed some type mismatches and removed unused vars. All handles now use Uint32_t in all function in glbapi.cpp

+++ src/gfxapi.cpp
@@ -1522,7 +1522,7 @@ GFX_StrPixelLen(
     
     for (loop = 0; loop < maxloop; loop++)
     {
-        outlen += infont->width[instr[loop]] + fontspacing;
+        outlen += infont->width[(int)instr[loop]] + fontspacing;
     }
     
     return outlen;
